v1.0.1
Features
1.Working Model Page that allows one to choose from 6 main Model ( which has proven that only the tinyllama Q6 quantization works ) -Additional Settings not implimented
2.Profile is improved with UI (Error 908 Present on the Model Test Phone)

Errors
Error 908
chat texbox is sticky at the bottom in the Model  test phone 
The App is not launching new sessions in termux so as to trigger the nwatcher.sh script...Ie it utilizes the same seesion breaking the functionality 

Notes
Contains Errors such as error 908 which has proven to be a pain .... I need to find a way to read from internal storage as reading from external storage is most devices is not an issue. The activity starter requesting for the persmision Read media storage does not work...next option is an extension that grants these permissions . If we lack options i will be forced to use in app assets and limit the user from selecting their prefered profile photo.

Feature Changes = make the select folder page more user friendly complete overhaul.